Improved order of Manual TOC
When I first started reading the documenation I found that I was going all over the place. Since the TOC is not alphabetically sorted, the TOC should be sorted in a way that the user can cover every page from top to bottom. Of course this is slightly subjective for some topics, but learning Game studio right after Getting started seems one of the first things you would want to cover.